Admin portal: Refactor AddEdit component
